requirements.txt
Django
djangorestframework
psycopg2
django-filter
django-cors-headers
python-decouple
gunicorn
whitenoise
boto3
drf-yasg
black
djangorestframework-simplejwt
xlsxwriter
pandas
openpyxl